pyproject.toml
[project]
name = "gazefy"
version = "0.1.0"
description = "Real-time screen monitoring and AI-driven automation for VDI-hosted applications"
requires-python = ">=3.11"
license = "Apache-2.0"

# Core deps — CI-safe, cross-platform
dependencies = [
    "numpy>=1.26",
    "opencv-python-headless>=4.9",
    "Pillow>=10.0",
    "pyyaml>=6.0",
    "httpx>=0.27",
]

[project.optional-dependencies]
# macOS platform deps — not installed in CI
platform = [
    "mss>=9.0",
    "pyautogui>=0.9",
    "pyobjc-framework-Quartz>=10",
]
# ML inference + training
ml = [
    "ultralytics>=8.3",
    "onnxruntime>=1.17",
]
# LLM providers
llm = [
    "anthropic>=0.40",
]
# Training-specific augmentations
training = [
    "albumentations>=1.4",
]
# Knowledge module (manual parsing)
knowledge = [
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
]
# Development tools
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]
# Everything for local development on macOS
all = [
    "gazefy[platform,ml,llm,training,dev]",
]

[project.scripts]
gazefy = "gazefy.cli:main"

[tool.setuptools.packages.find]
include = ["gazefy*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
